requirements.txt
Django==4.2.7
Pillow==10.0.0
django-crispy-forms==2.0
crispy-bootstrap5==0.7
django-allauth==0.54.0
django-filter==23.2
djangorestframework==3.14.0
markdown==3.4.4
django-cors-headers==4.2.0
python-dotenv==1.0.0
dj-database-url==2.1.0
psycopg2-binary==2.9.7
gunicorn==21.2.0
whitenoise==6.5.0

